requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
black==25.1.0
boto3==1.40.59
botocore==1.40.59
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
Django==5.2.8
django-cloudinary-storage==0.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
postgrest==2.23.2
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
realtime==2.23.2
requests==2.32.5
s3transfer==0.14.0
scapy==2.6.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.23.2
StrEnum==0.4.15
supabase==2.23.2
supabase-auth==2.23.2
supabase-functions==2.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0

# REST API Dependencies
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==23.5

# Production Dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
whitenoise==6.6.0
